+++ spec/spec_helper.rb
@@ -5,3 +5,9 @@ require "vagrant"
 require_relative "../lib/vagrant-triggers"
 
 VagrantPlugins::Triggers::Plugin.init_i18n
+
+RSpec.configure do |config|
+  config.mock_with :rspec do |mocks|
+    mocks.syntax = :should
+  end
+end